+++ src/__internal__/radio-button-mapper/radio-button-mapper.component.tsx
@@ -57,7 +57,6 @@ const RadioButtonMapper = ({
   const onChangeProp = useCallback(
     (event) => {
       onChange?.(event);
-      /* istanbul ignore else */
       if (!isControlled) {
         setCheckedValue(event.target.value);
       }