+++ src/components/checkbox/checkbox-test.stories.tsx
@@ -1,18 +1,12 @@
 import React, { useState } from "react";
 import { action } from "@storybook/addon-actions";
 
-import { Checkbox, CheckboxGroup, CheckboxGroupProps, CheckboxProps } from ".";
+import { Checkbox, CheckboxProps } from ".";
 import Box from "../box";
-import CarbonProvider from "../carbon-provider";
 
 export default {
   title: "Checkbox/Test",
-  includeStories: [
-    "Default",
-    "WithLongLabel",
-    "WithNewValidation",
-    "WithNewValidationGroup",
-  ],
+  includeStories: ["Default", "WithLongLabel"],
   parameters: {
     info: { disable: true },
     chromatic: {
@@ -116,43 +110,4 @@ WithLongLabel.storyName = "With long label";
 WithLongLabel.args = {
   label: "A really long description that will wrap onto the next line.",
   size: "",
-};
-
-export const WithNewValidation = (props: Partial<CheckboxProps>) => {
-  return (
-    <CarbonProvider validationRedesignOptIn>
-      <Checkbox label="Checkbox 1" {...props} />
-    </CarbonProvider>
-  );
-};
-
-WithNewValidation.args = {
-  error: "Error message",
-  warning: "",
-  fieldHelp: "field help text",
-  labelHelp: "label help text",
-  required: false,
-  checked: false,
-};
-
-export const WithNewValidationGroup = ({
-  ...props
-}: Partial<CheckboxGroupProps>) => {
-  return (
-    <CarbonProvider validationRedesignOptIn>
-      <CheckboxGroup legend="Checkbox legend" {...props}>
-        <Checkbox label="Checkbox 1" labelHelp="this shouldn't render" />
-        <Checkbox label="Checkbox 2" fieldHelp="this shouldn't render either" />
-        <Checkbox label="Checkbox 3" />
-      </CheckboxGroup>
-    </CarbonProvider>
-  );
-};
-
-WithNewValidationGroup.args = {
-  error: "Error message",
-  warning: "",
-  legendHelp: "Legend help text",
-  legendInline: false,
-  required: false,
 };